ajout de commentaire

+++ controllers/sauces.js
@@ -1,7 +1,8 @@
+//importation des packages
 const Sauce = require("../models/Sauce");
 const fs = require("fs");
 
-/////// CONTROLLER CREATE SAUCES
+// controller pour créer une sauce si authentification ok
 exports.createSauce = (req, res, next) => {
   const sauceObject = JSON.parse(req.body.sauce);
 
@@ -15,7 +16,6 @@ exports.createSauce = (req, res, next) => {
       req.file.filename
     }`,
   });
-
   sauce
     .save()
     .then(() => {
@@ -26,7 +26,7 @@ exports.createSauce = (req, res, next) => {
     });
 };
 
-/////// CONTROLLER ALL SAUCES
+// controller pour lire toutes les sauce  authentification ok
 exports.getAllSauces = (req, res, next) => {
   Sauce.find()
     .then((sauces) => {
@@ -37,14 +37,14 @@ exports.getAllSauces = (req, res, next) => {
     });
 };
 
-/////// CONTROLLER ONE SAUCE
+// controller pour lire une sauce
 exports.getOneSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => res.status(200).json(sauce))
     .catch((error) => res.status(404).json({ error }));
 };
 
-/////// CONTROLLER UPDATE ONE SAUCE
+// controller pour modifier une sauce si authentification ok
 exports.updateSauce = (req, res, next) => {
   const sauceObject = req.file
     ? {
@@ -56,46 +56,54 @@ exports.updateSauce = (req, res, next) => {
       }
     : { ...req.body };
 
+  delete sauceObject._id;
   delete sauceObject._userId;
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
       if (sauce.userId !== req.auth.userId) {
-        res.status(401).json({ message: "Modification non autorisé" });
+        res
+          .status(401)
+          .json({ message: "Mise a jour de lasauce non autorisée" });
       } else {
-        Sauce.findOne({ _id: req.params.id })
-          .then((sauce) => {
-            const filename = sauce.imageUrl.split("/images/")[1];
-            fs.unlink(`images/${filename}`, () => {
-              Sauce.updateOne(
-                { _id: req.params.id },
-                { ...sauceObject, _id: req.params.id }
+        if (req.file) {
+          const filename = sauce.imageUrl.split("/images/")[1];
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.updateOne(
+              { _id: req.params.id },
+              { ...sauceObject, _id: req.params.id }
+            )
+              .then(() =>
+                res.status(200).json({ message: "Sauce mise à jour" })
               )
-                .then(() =>
-                  res.status(200).json({ message: "Sauce mise a jour" })
-                )
-                .catch((error) => res.status(400).json({ error }));
-            });
-          })
-          .catch((error) => res.status(401).json({ error }));
+              .catch((error) => res.status(401).json({ error }));
+          });
+        } else {
+          Sauce.updateOne(
+            { _id: req.params.id },
+            { ...sauceObject, _id: req.params.id }
+          )
+            .then(() => res.status(200).json({ message: "Sauce mise à jour" }))
+            .catch((error) => res.status(401).json({ error }));
+        }
       }
     })
-    .catch((error) => res.status(500).json({ error }));
+    .catch((error) => res.status(400).json({ error }));
 };
 
-/////// CONTROLLER DELETE ONE SAUCE
+// controller pour supprimer une sauce si authentification ok
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
       if (sauce.userId !== req.auth.userId) {
-        return res.status(401).json({ error: "Suppression non autorisée !" });
+        return res.status(401).json({ message: "Suppression non autorisée !" });
       } else {
         const filename = sauce.imageUrl.split("/images/")[1];
         fs.unlink(`images/${filename}`, () => {
           Sauce.deleteOne({ _id: req.params.id })
             .then(() => res.status(200).json({ message: "Sauce supprimée" }))
-            .catch((error) => res.status(400).json({ error }));
+            .catch((error) => res.status(401).json({ error }));
         });
       }
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch((error) => res.status(500).json({ error }));
 };
